requirements.txt
gdown==3.12.2
matplotlib==3.3.1
nibabel==3.1.1
requests==2.24.0
scikit-image==0.17.2
segmentation-models-pytorch==0.1.0
statsmodels==0.12.0
